+++ packages/brick-utils/src/setProperties.ts
@@ -8,7 +8,7 @@ const replaceTemplateValue = (
 ): any => {
   const [, namespace, field, defaultValue, type] = matches;
   let result;
-  if (namespace === "query") {
+  if (namespace === "QUERY" || namespace === "query") {
     if (field === "*") {
       result = context.query;
     } else {
@@ -16,8 +16,10 @@ const replaceTemplateValue = (
         ? context.query.get(field)
         : defaultValue;
     }
-  } else if (namespace === "event") {
+  } else if (namespace === "EVENT" || namespace === "event") {
     result = field === "*" ? context.event : get(context.event, field);
+  } else if (namespace === "APP") {
+    result = get(context.app, field);
   } else if (context.match) {
     result = context.match.params[field];
     if (result === undefined) {
@@ -73,14 +75,14 @@ export const computeRealValue = (
     return newValue;
   }
   const matches = value.match(
-    /^\$\{(?:(query|event)\.)?([^|=}]+)(?:=([^|]*))?(?:\|(string|number|bool(?:ean)?|json))?\}$/
+    /^\$\{(?:(QUERY|EVENT|query|event|APP)\.)?([^|=}]+)(?:=([^|]*))?(?:\|(string|number|bool(?:ean)?|json))?\}$/
   );
   if (matches) {
     return replaceTemplateValue(matches, context);
   }
 
   return value.replace(
-    /\$\{(?:(query|event)\.)?([^|=}]+)(?:=([^|]*))?(?:\|(string|number|bool(?:ean)?|json))?\}/g,
+    /\$\{(?:(QUERY|EVENT|query|event|APP)\.)?([^|=}]+)(?:=([^|]*))?(?:\|(string|number|bool(?:ean)?|json))?\}/g,
     (match, query, field, defaultValue, type) =>
       replaceTemplateValue([match, query, field, defaultValue, type], context)
   );