+++ packages/brick-dll/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.12](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-dll@0.6.11...@easyops/brick-dll@0.6.12) (2019-09-17)
+
+**Note:** Version bump only for package @easyops/brick-dll
+
+## [0.6.11](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-dll@0.6.10...@easyops/brick-dll@0.6.11) (2019-09-17)
+
+**Note:** Version bump only for package @easyops/brick-dll
+
+## [0.6.10](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-dll@0.6.9...@easyops/brick-dll@0.6.10) (2019-09-16)
+
+**Note:** Version bump only for package @easyops/brick-dll
+
+## [0.6.9](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-dll@0.6.8...@easyops/brick-dll@0.6.9) (2019-09-16)
+
+**Note:** Version bump only for package @easyops/brick-dll
+
+## [0.6.8](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-dll@0.6.7...@easyops/brick-dll@0.6.8) (2019-09-16)
+
+**Note:** Version bump only for package @easyops/brick-dll
+
+## [0.6.7](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-dll@0.6.6...@easyops/brick-dll@0.6.7) (2019-09-14)
+
+**Note:** Version bump only for package @easyops/brick-dll
+
+## [0.6.6](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-dll@0.6.5...@easyops/brick-dll@0.6.6) (2019-09-11)
+
+**Note:** Version bump only for package @easyops/brick-dll
+
+## [0.6.5](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-dll@0.6.4...@easyops/brick-dll@0.6.5) (2019-09-10)
+
+**Note:** Version bump only for package @easyops/brick-dll
+
 ## [0.6.4](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-dll@0.6.3...@easyops/brick-dll@0.6.4) (2019-09-06)
 
 **Note:** Version bump only for package @easyops/brick-dll