+++ packages/brick-http/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.2](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-http@0.2.1...@easyops/brick-http@0.2.2) (2019-09-11)
+
+**Note:** Version bump only for package @easyops/brick-http
+
 ## [0.2.1](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-http@0.2.0...@easyops/brick-http@0.2.1) (2019-08-30)
 
 **Note:** Version bump only for package @easyops/brick-http