+++ packages/sdk-scripts/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.8.0](https://git.easyops.local/anyclouds/next-core/compare/@easyops/sdk-scripts@0.7.1...@easyops/sdk-scripts@0.8.0) (2019-09-11)
+
+### Features
+
+- move certain dev dependencies to root ([ab25738](https://git.easyops.local/anyclouds/next-core/commits/ab25738))
+
 ## [0.7.1](https://git.easyops.local/anyclouds/next-core/compare/@easyops/sdk-scripts@0.7.0...@easyops/sdk-scripts@0.7.1) (2019-08-30)
 
 **Note:** Version bump only for package @easyops/sdk-scripts